Revert "Refactor Loadtest script (#336)"
This reverts commit c8b83c701aefa707825b3f88c519518e8c7e311c.
