+++ loadtest/main.go
@@ -1,33 +1,138 @@
 package main
 
 import (
+	"context"
 	"encoding/json"
+	"flag"
 	"fmt"
 	"math/rand"
+	"os"
 	"os/exec"
+	"path/filepath"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
 
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	tokenfactorytypes "github.com/sei-protocol/sei-chain/x/tokenfactory/types"
 
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
-	"github.com/cosmos/cosmos-sdk/std"
 
-	"github.com/cosmos/cosmos-sdk/x/auth/tx"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 
+	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/codec/types"
+	"github.com/cosmos/cosmos-sdk/std"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	typestx "github.com/cosmos/cosmos-sdk/types/tx"
+	"github.com/cosmos/cosmos-sdk/x/auth/tx"
 	"github.com/sei-protocol/sei-chain/app"
+	"github.com/sei-protocol/sei-chain/utils"
 	dextypes "github.com/sei-protocol/sei-chain/x/dex/types"
+	"google.golang.org/grpc"
 )
+
+const (
+	Bank                 string = "bank"
+	FailureBankMalformed string = "failure_bank_malformed"
+	FailureBankInvalid   string = "failure_bank_invalid"
+	FailureDexMalformed  string = "failure_dex_malformed"
+	FailureDexInvalid    string = "failure_dex_invalid"
+	Dex                  string = "dex"
+	Staking              string = "staking"
+	Tokenfactory         string = "tokenfactory"
+	Limit                string = "limit"
+	Market               string = "market"
+)
+
+type EncodingConfig struct {
+	InterfaceRegistry types.InterfaceRegistry
+	// NOTE: this field will be renamed to Codec
+	Marshaler codec.Codec
+	TxConfig  client.TxConfig
+	Amino     *codec.LegacyAmino
+}
+
+type Config struct {
+	BatchSize         uint64                `json:"batch_size"`
+	ChainID           string                `json:"chain_id"`
+	OrdersPerBlock    uint64                `json:"orders_per_block"`
+	Rounds            uint64                `json:"rounds"`
+	MessageType       string                `json:"message_type"`
+	PriceDistr        NumericDistribution   `json:"price_distribution"`
+	QuantityDistr     NumericDistribution   `json:"quantity_distribution"`
+	DexMsgTypeDistr   MsgTypeDistribution   `json:"dex_message_type_distribution"`
+	ContractDistr     ContractDistributions `json:"contract_distribution"`
+	Constant          bool                  `json:"constant"`
+	ConstLoadInterval int64                 `json:"const_load_interval"`
+}
+
+type NumericDistribution struct {
+	Min         sdk.Dec `json:"min"`
+	Max         sdk.Dec `json:"max"`
+	NumDistinct int64   `json:"number_of_distinct_values"`
+}
+
+func (d *NumericDistribution) Sample() sdk.Dec {
+	steps := sdk.NewDec(rand.Int63n(d.NumDistinct))
+	return d.Min.Add(d.Max.Sub(d.Min).QuoInt64(d.NumDistinct).Mul(steps))
+}
+
+// Invalid numeric distribution sample
+func (d *NumericDistribution) InvalidSample() sdk.Dec {
+	steps := sdk.NewDec(rand.Int63n(d.NumDistinct))
+	if rand.Float64() < 0.5 {
+		return d.Min.Add(d.Max.Sub(d.Min).QuoInt64(d.NumDistinct).Mul(steps))
+	}
+	return d.Max.Add(d.Max.Sub(d.Min).QuoInt64(d.NumDistinct).Mul(steps))
+}
+
+type MsgTypeDistribution struct {
+	LimitOrderPct  sdk.Dec `json:"limit_order_percentage"`
+	MarketOrderPct sdk.Dec `json:"market_order_percentage"`
+}
+
+func (d *MsgTypeDistribution) Sample() string {
+	if !d.LimitOrderPct.Add(d.MarketOrderPct).Equal(sdk.OneDec()) {
+		panic("Distribution percentages must add up to 1")
+	}
+	randNum := sdk.MustNewDecFromStr(fmt.Sprintf("%f", rand.Float64()))
+	if randNum.LT(d.LimitOrderPct) {
+		return Limit
+	}
+	return "market"
+}
+
+type ContractDistributions []ContractDistribution
+
+func (d *ContractDistributions) Sample() string {
+	if !utils.Reduce(*d, func(i ContractDistribution, o sdk.Dec) sdk.Dec { return o.Add(i.Percentage) }, sdk.ZeroDec()).Equal(sdk.OneDec()) {
+		panic("Distribution percentages must add up to 1")
+	}
+	randNum := sdk.MustNewDecFromStr(fmt.Sprintf("%f", rand.Float64()))
+	cumPct := sdk.ZeroDec()
+	for _, dist := range *d {
+		cumPct = cumPct.Add(dist.Percentage)
+		if randNum.LTE(cumPct) {
+			return dist.ContractAddr
+		}
+	}
+	panic("this should never be triggered")
+}
+
+type ContractDistribution struct {
+	ContractAddr string  `json:"contract_address"`
+	Percentage   sdk.Dec `json:"percentage"`
+}
 
 var (
 	TestConfig         EncodingConfig
+	TxClient           typestx.ServiceClient
 	StakingQueryClient stakingtypes.QueryClient
+	TxHashFile         *os.File
+	ChainID            string
 	//// Staking specific variables
 	Validators []stakingtypes.Validator
 	// DelegationMap is a map of delegator -> validator -> delegated amount
@@ -59,31 +164,112 @@ func init() {
 	app.ModuleBasics.RegisterInterfaces(TestConfig.InterfaceRegistry)
 }
 
-func run() {
-	client := NewLoadTestClient()
-	client.SetValidators()
-	config := client.LoadTestConfig
-
-	defer client.Close()
-
-	if config.TxsPerBlock < config.MsgsPerTx {
-		panic("Must have more TxsPerBlock than MsgsPerTx")
+func run(config Config) {
+	ChainID = config.ChainID
+	grpcConn, _ := grpc.Dial(
+		"127.0.0.1:9090",
+		grpc.WithInsecure(),
+	)
+	defer grpcConn.Close()
+	TxClient = typestx.NewServiceClient(grpcConn)
+	StakingQueryClient = stakingtypes.NewQueryClient(grpcConn)
+	createOutputFiles()
+	var mu sync.Mutex
+	batchSize := config.BatchSize
+	if config.OrdersPerBlock < batchSize {
+		panic("Must have more orders per block than batch size")
 	}
+	setValidators(config)
+	DelegationMap = map[string]map[string]int{}
+	TokenFactoryDenomOwner = map[string]string{}
+	numberOfAccounts := config.OrdersPerBlock / batchSize * 2 // * 2 because we need two sets of accounts
+	activeAccounts := []int{}
+	inactiveAccounts := []int{}
+	for i := 0; i < int(numberOfAccounts); i++ {
+		if i%2 == 0 {
+			activeAccounts = append(activeAccounts, i)
+		} else {
+			inactiveAccounts = append(inactiveAccounts, i)
+		}
+	}
+	wgs := []*sync.WaitGroup{}
+	sendersList := [][]func() string{}
 
 	configString, _ := json.Marshal(config)
-	fmt.Printf("Running with \n %s \n", string(configString))
+	fmt.Printf("Running with \n %s \ns", string(configString))
 
 	fmt.Printf("%s - Starting block prepare\n", time.Now().Format("2006-01-02T15:04:05"))
-	workgroups, sendersList := client.BuildTxs()
+	for i := 0; i < int(config.Rounds); i++ {
+		fmt.Printf("Preparing %d-th round\n", i)
+		wg := &sync.WaitGroup{}
+		var senders []func() string
+		wgs = append(wgs, wg)
+		for j, account := range activeAccounts {
+			key := GetKey(uint64(account))
 
-	client.SendTxs(workgroups, sendersList)
+			msg, failureExpected := generateMessage(config, key, batchSize)
+			txBuilder := TestConfig.TxConfig.NewTxBuilder()
+			_ = txBuilder.SetMsgs(msg)
+			seqDelta := uint64(i / 2)
+			mode := typestx.BroadcastMode_BROADCAST_MODE_SYNC
+			if j == len(activeAccounts)-1 {
+				mode = typestx.BroadcastMode_BROADCAST_MODE_BLOCK
+			}
+			// Note: There is a potential race condition here with seqnos
+			// in which a later seqno is delievered before an earlier seqno
+			// In practice, we haven't run into this issue so we'll leave this
+			// as is.
+			sender := SendTx(key, &txBuilder, mode, seqDelta, &mu, failureExpected)
+			wg.Add(1)
+			senders = append(senders, func() string {
+				defer wg.Done()
+				return sender()
+			})
+		}
+		sendersList = append(sendersList, senders)
 
-	// Records the resulting TxHash to file
-	client.WriteTxHashToFile()
+		inactiveAccounts, activeAccounts = activeAccounts, inactiveAccounts
+	}
+
+	lastHeight := getLastHeight()
+	for i := 0; i < int(config.Rounds); i++ {
+		newHeight := getLastHeight()
+		for newHeight == lastHeight {
+			time.Sleep(10 * time.Millisecond)
+			newHeight = getLastHeight()
+		}
+		fmt.Printf("Sending %d-th block\n", i)
+		senders := sendersList[i]
+		wg := wgs[i]
+		for _, sender := range senders {
+			go sender()
+		}
+		wg.Wait()
+		lastHeight = newHeight
+	}
 	fmt.Printf("%s - Finished\n", time.Now().Format("2006-01-02T15:04:05"))
 }
 
-func generateMessage(config Config, key cryptotypes.PrivKey, msgPerTx uint64) (sdk.Msg, bool) {
+func createOutputFiles() {
+	userHomeDir, _ := os.UserHomeDir()
+	_ = os.Mkdir(filepath.Join(userHomeDir, "outputs"), os.ModePerm)
+	filename := filepath.Join(userHomeDir, "outputs", "test_tx_hash")
+	_ = os.Remove(filename)
+	file, _ := os.OpenFile(filename, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o644)
+	TxHashFile = file
+}
+
+func setValidators(config Config) {
+	if strings.Contains(config.MessageType, "staking") {
+		if resp, err := StakingQueryClient.Validators(context.Background(), &stakingtypes.QueryValidatorsRequest{}); err != nil {
+			panic(err)
+		} else {
+			Validators = resp.Validators
+		}
+	}
+}
+
+func generateMessage(config Config, key cryptotypes.PrivKey, batchSize uint64) (sdk.Msg, bool) {
 	var msg sdk.Msg
 	messageTypes := strings.Split(config.MessageType, ",")
 	rand.Seed(time.Now().UnixNano())
@@ -103,7 +289,7 @@ func generateMessage(config Config, key cryptotypes.PrivKey, msgPerTx uint64) (s
 		price := config.PriceDistr.Sample()
 		quantity := config.QuantityDistr.Sample()
 		contract := config.ContractDistr.Sample()
-		orderPlacements := generateDexOrderPlacements(config, key, msgPerTx, price, quantity)
+		orderPlacements := generateDexOrderPlacements(config, key, batchSize, price, quantity)
 		amount, err := sdk.ParseCoinsNormalized(fmt.Sprintf("%d%s", price.Mul(quantity).Ceil().RoundInt64(), "usei"))
 		if err != nil {
 			panic(err)
@@ -182,7 +368,7 @@ func generateMessage(config Config, key cryptotypes.PrivKey, msgPerTx uint64) (s
 		price := config.PriceDistr.InvalidSample()
 		quantity := config.QuantityDistr.InvalidSample()
 		contract := config.ContractDistr.Sample()
-		orderPlacements := generateDexOrderPlacements(config, key, msgPerTx, price, quantity)
+		orderPlacements := generateDexOrderPlacements(config, key, batchSize, price, quantity)
 		amount, err := sdk.ParseCoinsNormalized(fmt.Sprintf("%d%s", price.Mul(quantity).Ceil().RoundInt64(), "usei"))
 		if err != nil {
 			panic(err)
@@ -197,7 +383,7 @@ func generateMessage(config Config, key cryptotypes.PrivKey, msgPerTx uint64) (s
 		price := config.PriceDistr.Sample()
 		quantity := config.QuantityDistr.Sample()
 		contract := config.ContractDistr.Sample()
-		orderPlacements := generateDexOrderPlacements(config, key, msgPerTx, price, quantity)
+		orderPlacements := generateDexOrderPlacements(config, key, batchSize, price, quantity)
 		var amountUsei int64
 		if rand.Float64() < 0.5 {
 			amountUsei = 10000 * price.Mul(quantity).Ceil().RoundInt64()
@@ -231,7 +417,7 @@ func generateDexOrderPlacements(config Config, key cryptotypes.PrivKey, batchSiz
 	if config.MessageType == "failure_bank_malformed" {
 		orderType = -1
 	} else {
-		dexMsgType := config.MsgTypeDistr.Sample()
+		dexMsgType := config.DexMsgTypeDistr.Sample()
 		switch dexMsgType {
 		case Limit:
 			orderType = dextypes.OrderType_LIMIT
@@ -321,5 +507,15 @@ func getLastHeight() int {
 }
 
 func main() {
-	run()
+	flag.Parse()
+	config := Config{}
+	pwd, _ := os.Getwd()
+
+	fileName := "/loadtest/config.json"
+	file, _ := os.ReadFile(pwd + fileName)
+	if err := json.Unmarshal(file, &config); err != nil {
+		panic(err)
+	}
+
+	run(config)
 }
