+++ loadtest/sign.go
@@ -6,7 +6,6 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"sync"
 	"time"
 
 	"github.com/cosmos/cosmos-sdk/client"
@@ -27,43 +26,7 @@ type AccountInfo struct {
 	Mnemonic string `json:"mnemonic"`
 }
 
-type SignerInfo struct {
-	AccountNumber  uint64
-	SequenceNumber uint64
-	mutex          *sync.Mutex
-}
-
-func NewSignerInfo(accountNumber uint64, sequenceNumber uint64) *SignerInfo {
-	return &SignerInfo{
-		AccountNumber:  accountNumber,
-		SequenceNumber: sequenceNumber,
-		mutex:          &sync.Mutex{},
-	}
-}
-
-func (si *SignerInfo) IncrementAccountNumber() {
-	si.mutex.Lock()
-	defer si.mutex.Unlock()
-	si.AccountNumber++
-}
-
-type SignerClient struct {
-	CachedAccountSeqNum *sync.Map
-	CachedAccountKey    *sync.Map
-}
-
-func NewSignerClient() *SignerClient {
-	return &SignerClient{
-		CachedAccountSeqNum: &sync.Map{},
-		CachedAccountKey:    &sync.Map{},
-	}
-}
-
-func (sc *SignerClient) GetKey(accountIdx uint64) cryptotypes.PrivKey {
-	if val, ok := sc.CachedAccountKey.Load(accountIdx); ok {
-		privKey := val.(cryptotypes.PrivKey)
-		return privKey
-	}
+func GetKey(accountIdx uint64) cryptotypes.PrivKey {
 	userHomeDir, _ := os.UserHomeDir()
 	accountKeyFilePath := filepath.Join(userHomeDir, "test_accounts", fmt.Sprintf("ta%d.json", accountIdx))
 	jsonFile, err := os.Open(accountKeyFilePath)
@@ -85,19 +48,12 @@ func (sc *SignerClient) GetKey(accountIdx uint64) cryptotypes.PrivKey {
 	algo, _ := keyring.NewSigningAlgoFromString(algoStr, keyringAlgos)
 	hdpath := hd.CreateHDPath(sdk.GetConfig().GetCoinType(), 0, 0).String()
 	derivedPriv, _ := algo.Derive()(accountInfo.Mnemonic, "", hdpath)
-	privKey := algo.Generate()(derivedPriv)
-
-	// Cache this so we don't need to regenerate it
-	sc.CachedAccountKey.Store(accountIdx, privKey)
-	return privKey
+	return algo.Generate()(derivedPriv)
 }
 
-func (sc *SignerClient) SignTx(chainID string, txBuilder *client.TxBuilder, privKey cryptotypes.PrivKey, seqDelta uint64) {
+func SignTx(txBuilder *client.TxBuilder, privKey cryptotypes.PrivKey, seqDelta uint64) {
 	var sigsV2 []signing.SignatureV2
-	signerInfo := sc.GetAccountNumberSequenceNumber(privKey)
-	accountNum := signerInfo.AccountNumber
-	seqNum := signerInfo.SequenceNumber
-
+	accountNum, seqNum := GetAccountNumberSequenceNumber(privKey)
 	seqNum += seqDelta
 	sigV2 := signing.SignatureV2{
 		PubKey: privKey.PubKey(),
@@ -111,7 +67,7 @@ func (sc *SignerClient) SignTx(chainID string, txBuilder *client.TxBuilder, priv
 	_ = (*txBuilder).SetSignatures(sigsV2...)
 	sigsV2 = []signing.SignatureV2{}
 	signerData := xauthsigning.SignerData{
-		ChainID:       chainID,
+		ChainID:       ChainID,
 		AccountNumber: accountNum,
 		Sequence:      seqNum,
 	}
@@ -127,13 +83,7 @@ func (sc *SignerClient) SignTx(chainID string, txBuilder *client.TxBuilder, priv
 	_ = (*txBuilder).SetSignatures(sigsV2...)
 }
 
-func (sc *SignerClient) GetAccountNumberSequenceNumber(privKey cryptotypes.PrivKey) SignerInfo {
-	if val, ok := sc.CachedAccountSeqNum.Load(privKey); ok {
-		signerinfo := val.(SignerInfo)
-		signerinfo.IncrementAccountNumber()
-		return signerinfo
-	}
-
+func GetAccountNumberSequenceNumber(privKey cryptotypes.PrivKey) (uint64, uint64) {
 	hexAccount := privKey.PubKey().Address()
 	address, err := sdk.AccAddressFromHex(hexAccount.String())
 	if err != nil {
@@ -160,8 +110,5 @@ func (sc *SignerClient) GetAccountNumberSequenceNumber(privKey cryptotypes.PrivK
 			panic(err)
 		}
 	}
-
-	signerInfo := *NewSignerInfo(account, seq)
-	sc.CachedAccountSeqNum.Store(privKey, signerInfo)
-	return signerInfo
+	return account, seq
 }
