+++ loadtest/tx.go
@@ -3,6 +3,7 @@ package main
 import (
 	"context"
 	"fmt"
+	"sync"
 	"time"
 
 	"github.com/cosmos/cosmos-sdk/client"
@@ -17,17 +18,17 @@ func SendTx(
 	txBuilder *client.TxBuilder,
 	mode typestx.BroadcastMode,
 	seqDelta uint64,
+	mu *sync.Mutex,
 	failureExpected bool,
-	loadtestClient LoadTestClient,
-) func() string { // TODO: do we need to return string?
+) func() string {
 	(*txBuilder).SetGasLimit(200000000)
 	(*txBuilder).SetFeeAmount([]sdk.Coin{
 		sdk.NewCoin("usei", sdk.NewInt(10000000)),
 	})
-	loadtestClient.SignerClient.SignTx(loadtestClient.ChainID, txBuilder, key, seqDelta)
+	SignTx(txBuilder, key, seqDelta)
 	txBytes, _ := TestConfig.TxConfig.TxEncoder()((*txBuilder).GetTx())
 	return func() string {
-		grpcRes, err := loadtestClient.TxClient.BroadcastTx(
+		grpcRes, err := TxClient.BroadcastTx(
 			context.Background(),
 			&typestx.BroadcastTxRequest{
 				Mode:    mode,
@@ -45,7 +46,7 @@ func SendTx(
 			// retry after a second until either succeed or fail for some other reason
 			fmt.Printf("Mempool full\n")
 			time.Sleep(1 * time.Second)
-			grpcRes, err = loadtestClient.TxClient.BroadcastTx(
+			grpcRes, err = TxClient.BroadcastTx(
 				context.Background(),
 				&typestx.BroadcastTxRequest{
 					Mode:    mode,
@@ -63,7 +64,11 @@ func SendTx(
 		if grpcRes.TxResponse.Code != 0 {
 			fmt.Printf("Error: %d, %s\n", grpcRes.TxResponse.Code, grpcRes.TxResponse.RawLog)
 		} else {
-			loadtestClient.AppendTxHash(grpcRes.TxResponse.TxHash)
+			mu.Lock()
+			defer mu.Unlock()
+			if _, err := TxHashFile.WriteString(fmt.Sprintf("%s\n", grpcRes.TxResponse.TxHash)); err != nil {
+				panic(err)
+			}
 			return grpcRes.TxResponse.TxHash
 		}
 		return ""
